+++ src/main/java/edu/csula/datascience/elasticsearch/ExportESApp.java
@@ -9,12 +9,12 @@ public class ExportESApp {
 	public static void main(String[] args) throws URISyntaxException {
 
 		//pass your cluster name to the constructor
-		MovieExporter movieExp = new MovieExporter("darkserith");
-		movieExp.exportToES();
-		List<Movie> movies = movieExp.getMovies();
+		MongoMovieImporter movieImporter = new MongoMovieImporter();
+		movieImporter.importMovies();
+		List<Movie> movies = movieImporter.getMovies();
 		
 		//pass in your cluster name to the constructor
-		TweetExporter tweetExp = new TweetExporter("darkserith", movies);
-		tweetExp.exportToES();
+		MovieExporter movieExp = new MovieExporter("darkserith", movies, 400);
+		movieExp.exportToES();
 	}	
 }
